@@ -1,4 +1,4 @@
-// swift-tools-version: 5.6
+// swift-tools-version: 5.9
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -20,10 +20,34 @@ let package = Package(
             dependencies: [
                 .product(name: "ArgumentParser", package: "swift-argument-parser")
             ],
-            path: "Sources"),
+            path: "Sources",
+            swiftSettings: [
+                            .enableExperimentalFeature("ConciseMagicFile"),
+                            .enableExperimentalFeature("ForwardTrailingClosures"),
+                            .enableExperimentalFeature("BareSlashRegexLiterals"),
+                            .enableExperimentalFeature("DeprecateApplicationMain"),
+                            .enableExperimentalFeature("ImportObjcForwardDeclarations"),
+                            .enableExperimentalFeature("DisableOutwardActorInference"),
+                            .enableExperimentalFeature("InternalImportsByDefault"),
+                            .enableExperimentalFeature("ExistentialAny"),
+                            .enableExperimentalFeature("ImplicitOpenExistentials"),
+                            .enableExperimentalFeature("StrictConcurrency")
+                        ]),
         .testTarget(
             name: "ToggleCipherTests",
             dependencies: ["ToggleCipher"],
-            path: "Tests")
+            path: "Tests",
+            swiftSettings: [
+                            .enableExperimentalFeature("ConciseMagicFile"),
+                            .enableExperimentalFeature("ForwardTrailingClosures"),
+                            .enableExperimentalFeature("BareSlashRegexLiterals"),
+                            .enableExperimentalFeature("DeprecateApplicationMain"),
+                            .enableExperimentalFeature("ImportObjcForwardDeclarations"),
+                            .enableExperimentalFeature("DisableOutwardActorInference"),
+                            .enableExperimentalFeature("InternalImportsByDefault"),
+                            .enableExperimentalFeature("ExistentialAny"),
+                            .enableExperimentalFeature("ImplicitOpenExistentials"),
+                            .enableExperimentalFeature("StrictConcurrency")
+                        ])
     ]
 )
