@@ -3,7 +3,7 @@
 import Foundation
 
 /// Supported types for `Object`
-public enum ObjectSupportedType {
+public enum ObjectSupportedType: Sendable {
     case bool(Bool)
     case int(Int)
     case number(Double)
